added configurable particle effects added new module editor for particle effects added ability to camera zoom when gui opened fixed camera moving when gui opened added CheckBox, CompoBox gui objects InputBox improvements Chat improvements Scroll improvements Gui rework font rendering improvements SoundLoader improvements TextureLoader improvements

+++ core/client/src/main/java/net/bfsr/client/renderer/font/IStringXOffsetSupplier.java
@@ -2,5 +2,5 @@ package net.bfsr.client.renderer.font;
 
 @FunctionalInterface
 public interface IStringXOffsetSupplier {
-    float get(String string, StringCache stringCache);
+    int get(String string, StringCache stringCache);
 }
